+++ ft_printf_dir/unsigned_base.c
@@ -6,13 +6,13 @@
 /*   By: ydag <ydag@student.codam.nl>                 +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2019/11/30 14:10:39 by ydag           #+#    #+#                */
-/*   Updated: 2019/12/02 13:04:16 by ydag          ########   odam.nl         */
+/*   Updated: 2019/12/20 15:36:29 by ydag          ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-void	unsigned_base(unsigned int n, int base, int up)
+void	unsigned_base(unsigned long long n, int base, int up)
 {
 	const char			upc[] = "0123456789ABCDEF";
 	const char			low[] = "0123456789abcdef";
@@ -22,7 +22,7 @@ void	unsigned_base(unsigned int n, int base, int up)
 		ptr = upc;
 	else
 		ptr = low;
-	if (n > (unsigned int)(base - 1))
+	if (n > (unsigned long long)(base - 1))
 		unsigned_base(n / base, base, up);
 	ft_putchar(ptr[(n % base)]);
 }
